% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.esonar.R
\name{locate.esonar}
\alias{locate.esonar}
\alias{locate.esonar.default}
\alias{locate.esonar.scsset}
\title{Locate Star Oddi Probe Data}
\usage{
locate.esonar(x, ...)

\method{locate}{esonar.default}(x, year, tow.id, remove = c("test", "use raw"), ...)

\method{locate}{esonar.scsset}(x, ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Study or survey year.}

\item{tow.id}{Character string(s) specifying a snow crab survey tow identifier(s) (e.g. \sQuote{GP354F}).}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}

\item{probe}{Character string specifying probe type or location. For example \sQuote{headline} specifies the Star Oddi 
probe attached to the trawl headline on snow crab surveys.}
}
\description{
Functions to locate Star Oddi probe data from various projects and surveys.
}
\section{Methods (by class)}{
\itemize{
\item \code{esonar.default}: Default method for locating eSonar data files.

\item \code{esonar.scsset}: \code{scsset} method for locating eSonar data files.
}}

\examples{
locate.esonar(2020) # Find snow crab survey eSonar data from 2020.
locate.esonar(2020, tow.id = "GP354F")
}
