% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.netmind.R
\name{locate.netmind}
\alias{locate.netmind}
\alias{locate.netmind.default}
\title{Locate Netmind Trawl Acoustic Data}
\usage{
locate.netmind(x, ...)

\method{locate}{netmind.default}(x, year, tow.id, remove = c("reject", "test", "invalid"), ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Study or survey year.}

\item{tow.id}{Character string(s) specifying a snow crab survey tow identifier(s) (e.g. \sQuote{GP354F}).}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}
}
\description{
Functions to locate Netmind probe data from different projects and surveys.
}
\section{Methods (by class)}{
\itemize{
\item \code{netmind.default}: Default method for locating Netmind acoustic data files.
}}

\examples{
# Global searches:
locate.netmind()     # Find all available Netmind data files.
locate.netmind(1999) # Find Netmind data files for the 1990 snow crab survey.
locate.netmind(1999:2004) # Find Netmind data files for the 1990-1994 snow crab survey.

# Specific searches:
locate.netmind(1999, tow.id = 100)
locate.netmind(1999, tow.id = "100")
locate.netmind(1999, tow.id = "S90100")
locate.netmind(tow.id = "355")
}
