% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.star.oddi.R
\name{locate.star.oddi}
\alias{locate.star.oddi}
\alias{locate.star.oddi.default}
\alias{locate.star.oddi.scs}
\alias{locate.star.oddi.scsset}
\alias{locate.star.oddi.nss}
\title{Locate Star Oddi Probe Data}
\usage{
locate.star.oddi(x, ...)

\method{locate}{star.oddi.default}(x, project = "scs", remove, ...)

\method{locate}{star.oddi.scs}(x, year, tow.id, probe = c("headline", "footrope", "tilt"), remove, ...)

\method{locate}{star.oddi.scsset}(x, ...)

\method{locate}{star.oddi.nss}(x, year, remove, ...)
}
\arguments{
\item{x}{Data object.}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}

\item{year}{Study or survey year.}

\item{tow.id}{Character string(s) specifying a snow crab survey tow identifier(s) (e.g. \sQuote{GP354F}).}

\item{probe}{Character string specifying probe type or location. For example \sQuote{headline} specifies the Star Oddi 
probe attached to the trawl headline on snow crab surveys.}
}
\description{
Functions to locate Star Oddi probe data from various projects and surveys.
}
\section{Methods (by class)}{
\itemize{
\item \code{star.oddi.default}: Default method for locating Star Oddi probe data files.

\item \code{star.oddi.scs}: Locate Star Oddi data files from the snow crab survey.

\item \code{star.oddi.scsset}: \code{scsset} method for locating Star Oddi data files.

\item \code{star.oddi.nss}: Locate Star Oddi data files from the Northumberland Strait survey.
}}

\examples{
locate.star.oddi(2020, probe = "headline") 
locate.star.oddi(2020, probe = "tilt", tow.id = "GP354F")
}
