% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.minilog.header.R
\name{read.minilog.header}
\alias{read.minilog.header}
\title{Read Minilog Header Information}
\usage{
\method{read.minilog}{header}(x, file, ...)
}
\arguments{
\item{x, file}{Minilog file names.}
}
\description{
Function that reads header information from Minilog data files.
}
\examples{
:
files <- locate.minilog(year = 1999, project = "scs")
read.minilog.header(files)

}
