% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.minilog.R
\name{read.minilog}
\alias{read.minilog}
\title{Read Depth-temperature Probe Data}
\usage{
read.minilog(x, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{file}{File name(s).}

\item{year}{Survey year(s).}

\item{survey}{Survey type, as determined by the \link{survey.scsset} function.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}
}
\description{
Function for reading VEMCO Minilog depth-temperature probe data. This function reads the ASCII format Minilog data. 
             The file's header information is stored in the object's attributes, and maybe extracted using the 
             'header' function. If multiple file names are specified, the data are appended together and columns are 
             added which contain the header information from each file.
}
\examples{
# Read snow crab survey Minilog data from 1997:
x <- read.minilog(1997)
}
