% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.minilog.R
\name{locate.minilog}
\alias{locate.minilog}
\alias{locate.minilog.default}
\title{Locate Minilog Trawl Acoustic Data}
\usage{
locate.minilog(x, ...)

\method{locate}{minilog.default}(x, year, tow.id, remove = c("reject", "test", "invalid"), ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Study or survey year.}

\item{tow.id}{Character string(s) specifying a snow crab survey tow identifier(s) (e.g. \sQuote{GP354F}).}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}
}
\description{
Functions to locate Minilog probe data from different projects and surveys.
}
\section{Methods (by class)}{
\itemize{
\item \code{minilog.default}: Default method for locating Minilog acoustic data files.
}}

\examples{
# Global searches:
locate.minilog()     # Find all available Minilog data files.
locate.minilog(1990) # Find Minilog data files for the 1990 snow crab survey.
locate.minilog(1990:1994) # Find Minilog data files for the 1990-1994 snow crab survey.

# Specific searches:
locate.minilog(1990, tow.id = 100)
locate.minilog(1990, tow.id = "100")
locate.minilog(1990, tow.id = "S90100")
locate.minilog(tow.id = "S90100")
}
